Create function to create conenction to db. Create function to get mpam due case data
